+++ src/test/java/unit/market/TestExchangeInfo.java
@@ -73,9 +73,45 @@ public class TestExchangeInfo {
     }
 
     @Test
+    public void testExchangeInfoWithOnePermission() {
+        String path = String.format("/api/v3/exchangeInfo?permissions=%s",
+                                    UrlBuilder.urlEncode("[\"MARGIN\"]"));
+        LinkedHashMap<String, Object> parameters = new LinkedHashMap<>();
+        ArrayList<String> permissions = new ArrayList<>();
+        permissions.add("MARGIN");
+        parameters.put("permissions", permissions);
+
+        Dispatcher dispatcher = MockWebServerDispatcher.getDispatcher(MockData.PREFIX, path, MockData.MOCK_RESPONSE, HttpMethod.GET, MockData.HTTP_STATUS_OK);
+        mockWebServer.setDispatcher(dispatcher);
+
+        SpotClientImpl client = new SpotClientImpl(null, null, baseUrl);
+        String result = client.createMarket().exchangeInfo(parameters);
+        assertEquals(MockData.MOCK_RESPONSE, result);
+    }
+
+    @Test
+    public void testExchangeInfoWithMultiplePermissions() {
+        String path = String.format("/api/v3/exchangeInfo?permissions=%s",
+                                    UrlBuilder.urlEncode("[\"MARGIN\",\"SPOT\"]"));
+        LinkedHashMap<String, Object> parameters = new LinkedHashMap<>();
+        ArrayList<String> permissions = new ArrayList<>();
+        permissions.add("MARGIN");
+        permissions.add("SPOT");
+        parameters.put("permissions", permissions);
+
+        Dispatcher dispatcher = MockWebServerDispatcher.getDispatcher(MockData.PREFIX, path, MockData.MOCK_RESPONSE, HttpMethod.GET, MockData.HTTP_STATUS_OK);
+        mockWebServer.setDispatcher(dispatcher);
+
+        SpotClientImpl client = new SpotClientImpl(null, null, baseUrl);
+        String result = client.createMarket().exchangeInfo(parameters);
+        assertEquals(MockData.MOCK_RESPONSE, result);
+    }
+
+
+    @Test
     public void testExchangeInfoWithInvalidType() {
         String path = String.format("/api/v3/exchangeInfo?symbols=%s",
-                UrlBuilder.urlEncode("[\"BNBUSDT\",\"BTCUSDT\"]"));
+                                    UrlBuilder.urlEncode("[\"BNBUSDT\",\"BTCUSDT\"]"));
         LinkedHashMap<String, Object> parameters = new LinkedHashMap<>();
         String[] symbols = {"BNBUSDT", "BTCUSDT"};
         parameters.put("symbols", symbols);
@@ -88,12 +124,50 @@ public class TestExchangeInfo {
     }
 
     @Test
+    public void testExchangeInfoPermissionsWithInvalidType() {
+        String path = String.format("/api/v3/exchangeInfo?permissions=%s",
+                                    UrlBuilder.urlEncode("[\"MARGIN\",\"SPOT\"]"));
+        LinkedHashMap<String, Object> parameters = new LinkedHashMap<>();
+        String[] permissions = {"MARGIN", "SPOT"};
+        parameters.put("permissions", permissions);
+
+        Dispatcher dispatcher = MockWebServerDispatcher.getDispatcher(MockData.PREFIX, path, MockData.MOCK_RESPONSE, HttpMethod.GET, MockData.HTTP_STATUS_OK);
+        mockWebServer.setDispatcher(dispatcher);
+
+        SpotClientImpl client = new SpotClientImpl(null, null, baseUrl);
+        assertThrows(BinanceConnectorException.class, () -> client.createMarket().exchangeInfo(parameters));
+    }
+
+    @Test
     public void testExchangeInfoWithDoubleParameter() {
         String path = String.format("/api/v3/exchangeInfo?symbols=%s",
-                UrlBuilder.urlEncode("[\"BNBUSDT\",\"BTCUSDT\"]"));
+                    UrlBuilder.urlEncode("[\"BNBUSDT\",\"BTCUSDT\"]"));
         LinkedHashMap<String, Object> parameters = new LinkedHashMap<>();
-        String[] symbols = {"BNBUSDT", "BTCUSDT"};
+        ArrayList<String> symbols = new ArrayList<>();
+        symbols.add("BNBUSDT");
+        symbols.add("BTCUSDT");
+
         parameters.put("symbols", symbols);
+        parameters.put("symbol", "ETHUSDT");
+
+        Dispatcher dispatcher = MockWebServerDispatcher.getDispatcher(MockData.PREFIX, path, MockData.MOCK_RESPONSE, HttpMethod.GET, MockData.HTTP_STATUS_OK);
+        mockWebServer.setDispatcher(dispatcher);
+
+        SpotClientImpl client = new SpotClientImpl(null, null, baseUrl);
+        assertThrows(BinanceConnectorException.class, () -> client.createMarket().exchangeInfo(parameters));
+    }
+
+    @Test
+    public void testExchangeInfoWithSymbolAndPermissionsParameters() {
+        String path = String.format("/api/v3/exchangeInfo?symbol=%s",
+                UrlBuilder.urlEncode("[\"ETHUSDT\"]"));
+        LinkedHashMap<String, Object> parameters = new LinkedHashMap<>();
+        
+        ArrayList<String> permissions = new ArrayList<>();
+        permissions.add("MARGIN");
+        permissions.add("SPOT");
+
+        parameters.put("permissions", permissions);
         parameters.put("symbol", "ETHUSDT");
 
         Dispatcher dispatcher = MockWebServerDispatcher.getDispatcher(MockData.PREFIX, path, MockData.MOCK_RESPONSE, HttpMethod.GET, MockData.HTTP_STATUS_OK);