+++ CHANGELOG.md
@@ -1,5 +1,27 @@
 # Changelog
 
+## 1.10.0 - 2022-10-17
+
+### Added
+- New endpoints for Crypto Loan:
+  - `POST /sapi/v1/loan/borrow` to borrow Crypto Loan 
+  - `GET /sapi/v1/loan/borrow/history` to get Loan Borrow History
+  - `GET/sapi/v1/loan/ongoing/orders` to get Loan Ongoing Orders
+  - `POST/sapi/v1/loan/repay` to repay Crypto Loan 
+  - `GET/sapi/v1/loan/repay/history` to get Loan Repayment History
+  - `POST/sapi/v1/loan/adjust/ltv` to adjust Crypto Loan LTV (Loan-To-Value)
+  - `GET/sapi/v1/loan/ltv/adjustment/history` to get Loan LTV (Loan-To-Value) Adjustment History
+
+### Changed
+- Changes to `GET /api/v3/exchangeInfo`:
+  - New optional parameter `permissions` added to display all symbols with the permissions matching the parameter provided. (eg.SPOT, MARGIN, LEVERAGED)
+  - If not provided, the default value will be `["SPOT","MARGIN","LEVERAGED"]`
+  - Cannot be combined with symbol or symbols
+- Updated `checkstyle.xml` to also include Indentation verification 
+
+### Fixed
+- `DecimalFormat` with `Locale.ENGLISH` for requests with `Double` to have `.` as decimal separator.
+
 ## 1.9.0 - 2022-08-30
 
 ### Added
@@ -126,7 +148,7 @@
   - `GET /sapi/v1/mining/payment/uid` to get Mining account earning.
 - New endpoint for BSwap:
 
-  - `GET /sapi/v1/bswap/unclaimedRewards `to get unclaimed rewards record.
+  - `GET /sapi/v1/bswap/unclaimedRewards` to get unclaimed rewards record.
   - `POST /sapi/v1/bswap/claimRewards` to claim swap rewards or liquidity rewards.
   - `GET /sapi/v1/bswap/claimedHistory` to get history of claimed rewards.